+++ conf/access_list.php
@@ -1,0 +1,7 @@
+<?php return array (
+  'allow' => array(),
+  'deny' => array(),
+);
+/* This is the global IP address black/whitelist. Add IP addresses or CIDR blocks
+** separated by a comma in quotes, for example: '127.0.0.1', '192.168.0.0/24'
+*/ ?>